The Register() action with no parameters simply returns a blank registration page to the user. The second Register() action accepts RegisterViewModel object as its parameter. Inside, the code checks whether the model contains valid data or not using the IsValid property of ModelState object. If the model is valid we create a new MyIdentityUser object and assign its UserName, Email, FullName and BirthDate properties. Then the code calls the CreateAsync() method of the UserManager class in an attempt to create a user account. The password is passed to the CreateAsync() method. Since the CreateAsync() is an asynchronous call, we block the execution by calling the Result property. The CreateAsync() method returns IdentityResult - an object that tells us whether the user creation was successful or not.
If the user creation succeeds (result.Succeeded property) we proceed to create the roles required by the application. The role creation is a one time operation. It should happen only if a role doesn't exist already. So, the code uses RoleManager to check whether a role already exists or not. This is done using RoleExistsAsync() method and by passing role name as its parameter.

If a role doesn't exist already, MyIdentityRole object representing that role is created. Its RoleName and Description properties are assigned the required values (NormalUser is the role name in this example). Then CreateAsync() is called on RoleManager in an attempt to create the role. If the role creation succeeds (Succeeded property returns true) we add the newly created user to the NormalUser role. This is done using AddToRoleAsync() method of UserManager.